--- example.env
@@ -1,5 +1,0 @@
-HOSTNAME=
-PORT=
-USERNAME=
-PASSWORD=
-OPENAI_API_KEY=
